Docs on settings up Visual Studio 2017 & CMake 2.7
There are a few things you need to do with Visual Studio 2017 as far as
CMake settings so that is documented now. Also Clion and Visual Studio
both ship with CMake 2.7 which the requirement was 2.8. There is no code
that breaks in the cmake files of used features so 2.7 is acceptable.
Also the Visual C++ compiler has different way to get the relative path
so there are now ifdefs for them.