Removed SDL2_ttf
Removed all known references to SDL2_ttf.
Replaced text.cpp and text.h contents with framework for new bitmap font format.

Added includes for SDL2_net
Forgot to add those before.

A definition was added for realpath() using something to do with MSVC.
No alternative was defined, and now it cannot build in Code::Blocks.
A cross-platform alternative is needed; it shouldn't be dependant on MSVC.