Implement an audio portability API
One major issue I've always had with brux-gdk is that it's not very portable since it has a strict dependency on SDL2.

This commit implements an audio portability API that allows non-SDL2 backends to be implemented, and improves the test suite included in the rte directory.

In case it's still needed though, the old version can still be found in test_old.

In the future, I want to also implement a portability API for graphics, input, and filesystem operations.

By the way, I also fixed some bugs in this commit ;)

Signed-off-by: hexaheximal <hexaheximal@proton.me>