Implement a function-based loop system
Previously, anything using Brux GDK would need to create a main loop which blocks the main thread.

Normally that's perfectly fine, but in a web browser blocking the main thread is something you should NEVER do.

The solution to this problem is to have the engine itself take care of the loop properly, and then the game just defines what should be run on each frame.