Fix getOS()
When running Brux GDK on postmarketOS, which uses Linux but not
glibc (and thus fails the __gnu_linux__ ifdef check), it crashes
because the xyGetOS() function is missing a return statement.

The most obvious solution would be to simply add a fallback, and
in fact an earlier version of the pull request fixing this
function did precisely that, but then I realized that the problem
is much more complicated.

Even though adding a fallback works, it misses the bigger
picture: The entire function is done incorrectly.

The Brux GDK documentation states that getOS() returns a string.
However, in practice, it does not. The commit which replaces
Miniswig with simplesquirrel accidentally binded getOS to
xyGetOS, instead of binding it to a function which returns the
output of SDL_GetPlatform() like it previously did.

Luckily, xyGetOS isn't used anywhere else, and the numerical OS
definitions appear to be completely unused. Because of that, this
commit simply turns xyGetOS into a wrapper for SDL_GetPlatform
and removes all references to the old numerical value system.